# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Snapcert Swagger Documentation",
        "version": "1.0.0",
        "title": "Tasks API",
        "contact": {
            "email": "sanket.raiturkar@snapperfuturetech.com"
        }
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3000",
    "basePath": "/api/v1",
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-api-token",
            "in": "header"
        }
    },
    "paths": {
...
        "/transaction/getCredById": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get CredByID",
                "description": "Get Transaction by ID",
                "produces": [
                    "application/json"
                ],
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "task id that needs to be deleted",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/GetCredByID"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                        
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/corporate/register": {
            "post": {
                "tags": [
                    "Corporate"
                ],
                "summary": "Corporate register",
                "description": "Corporate register",
                "produces": [
                    "application/json"
                ],
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "task id that needs to be deleted",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/Corporateregister"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                        
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/user/signin" : {
            "post" : {
                "tags": [
                    "User"
                  ],
                "summary" : "User Login",
                "description": "User Login ",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            
                                    "$ref": "#/definitions/User"
                                
                            
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login Successful"
                      
                       
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/user/resetpassword" : {
            "post" : {
                "tags": [
                    "User"
                  ],
               
                "summary" : "Reset Pasword",
                "description": "Reset Pasword",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "security": [
                    {
                      "ApiKeyAuth": []
                    }
                  ], 
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            
                                    "$ref": "#/definitions/resetpassword"
                                
                            
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reset Pasword Successful"
                      
                       
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/user/token_validate" : {
            "post" : {
                "tags": [
                    "User"
                  ],
               
                "summary" : "Token Validate",
                "description": "Token Validate",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "security": [
                    {
                      "ApiKeyAuth": []
                    }
                  ], 
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "tokenvalidate object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            
                                    "$ref": "#/definitions/tokenvalidate"
                                
                            
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "recaptcha passed"
                      
                       
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/user/forgotpassword" : {
            "post" : {
                "tags": [
                    "User"
                  ],
               
                "summary" : "Forgotpassword",
                "description": "Forgotpassword",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "security": [
                    {
                      "ApiKeyAuth": []
                    }
                  ], 
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "forgotpassword object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            
                                    "$ref": "#/definitions/forgotpassword"
                                
                            
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "forgotpassword Successfully"
                      
                       
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/user/token" : {
            "post" : {
                "tags": [
                    "User"
                  ],
               
                "summary" : "token",
                "description": "token",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "security": [
                    {
                      "ApiKeyAuth": []
                    }
                  ], 
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "token object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            
                                    "$ref": "#/definitions/token"
                                
                            
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                      
                       
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/user/resendOtpCall" : {
            "post" : {
                "tags": [
                    "User"
                  ],
               
                "summary" : "resendOtpCall",
                "description": "resendOtpCall",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "security": [
                    {
                      "ApiKeyAuth": []
                    }
                  ], 
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "resendOtpCall object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            
                                    "$ref": "#/definitions/resendOtpCall"
                                
                            
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                      
                       
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/user/otpShowHide" : {
            "post" : {
                "tags": [
                    "User"
                  ],
               
                "summary" : "otpShowHide",
                "description": "otpShowHide",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "security": [
                    {
                      "ApiKeyAuth": []
                    }
                  ], 
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "resendOtpCall object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            
                                    "$ref": "#/definitions/resendOtpCall"
                                
                            
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                      
                       
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
        "/organization/register" : {
            "post" : {
                "tags": [
                    "Organization"
                  ],
               
                "summary" : "Organization Register",
                "description": "Organization Register",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "OrganizationRegister object",
                        "required": true,
                        "schema": {
                            "type": "object",
                            
                                    "$ref": "#/definitions/OrganizationRegister"
                                
                            
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organization Register successfully"
                      
                       
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "resetpassword": {       
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "confirmPassword": {
                    "type": "string"
                }
            }
        },
        "tokenvalidate": {       
            "type": "object",
            "properties": {
                "recaptcha": {
                    "type": "string"
                }
            }
        },
...
        "forgotpassword": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "token": {
            "type": "object",
            "properties": {
                "session": {
                    "type": "object",

                    "properties": {
                    "userId": {
                        "type": "string"
                    },
                    "_id": {
                        "type": "string"
                    }  
                }               
                   
                  }
            }
        },
...
        "resendOtpCall": {
            "type": "object",
            "properties": {

                    "userId": {
                        "type": "string"
                    }    
                   
                
            }
        },
...
        "OrganizationRegister": {
            "type": "object",
            "properties": {              
                                
               "organizationAdmin": {
                    "type": "object",

                    "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    } ,
                    "type": {
                        "type": "string"
                    } ,
                    "email": {
                        "type": "string"
                    } ,
                    "phoneNumber": {
                        "type": "string"
                    } ,
                    "timeZone": {
                        "type": "string"
                    },
                    "createdBy": {
    
                        "type": "object",
    
                        "properties": {
    
                        "firstName": {
                            "type": "string"
                        },
                        "lastName": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        }
                    }                  
                       
                      },
                      "updatedBy": {
                        "type": "object",
    
                        "properties": {
                        "firstName": {
                            "type": "string"
                        },
                        "lastName": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        }    
                    }               
                       
                      }      
                }               
                   
                  }
            }
        },       
...
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}